refactor menu actions
