+++ src/electron/menu/menu-actions.js
@@ -1,0 +1,9 @@
+/**
+ * @author Marco Goebel
+ */
+
+module.exports = {
+  OPEN_CLIENT: "open-client",
+  EXPORT_XML: "export-xml",
+  SHOW_REPO: "show-repo"
+};
